+++ src/core/connection.ts
@@ -1,10 +1,13 @@
+import { ApiPath } from '@/constants';
 import { Collection } from '@/core/collection';
 import type { CollectionBase } from '@/interfaces/core';
 import type { FileUploadConfig, URLUploadConfig } from '@/types/collection';
-import type { CollectionResponse } from '@/types/response';
+import type { CollectionResponse, GetCollections } from '@/types/response';
 import { fromSnakeToCamel } from '@/utils';
 import { HttpClient } from '@/utils/httpClient';
 import { uploadToServer } from '@/utils/upload';
+
+const { collection } = ApiPath;
 
 class VdbHttpClient extends HttpClient {
   constructor(baseURL: string, apiKey: string) {
@@ -31,10 +34,68 @@ export class Connection {
    * else returns the default collection.
    */
   public async getCollection(id = 'default'): Promise<Collection> {
-    const res = await this.vhttp.get<CollectionResponse>(['collection', id]);
+    const res = await this.vhttp.get<CollectionResponse>([collection, id]);
     const convertedData = fromSnakeToCamel(res.data) as CollectionBase;
-    const collection = new Collection(this.vhttp, convertedData);
-    return collection;
+    const _collection = new Collection(this.vhttp, convertedData);
+    return _collection;
+  }
+
+  /**
+   * Get all Collections from db
+   * @returns
+   * Returns an array of Collection objects 
+   */
+  public async getCollections(): Promise<Collection[]> {
+    const res = await this.vhttp.get<GetCollections>([collection]);
+    const collectionsResponse = fromSnakeToCamel(res.data);
+    const collections = collectionsResponse.collections.map(_collection => {
+      const convertedData = fromSnakeToCamel(_collection) as CollectionBase;
+      return new Collection(this.vhttp, convertedData);
+    });
+    return collections;
+  }
+
+  /**
+   * Create a new collection
+   * @param name - Name of the collection
+   * @param description - Description of the collection
+   * @returns
+   * Returns a new Collection object
+   */
+  public createCollection = async (name: string, description: string) => {
+    const res = await this.vhttp.post<CollectionResponse, object>(
+      [collection],
+      {
+        name,
+        description,
+      }
+    );
+    const convertedData = fromSnakeToCamel(res.data) as CollectionBase;
+    return new Collection(this.vhttp, convertedData);
+  };
+
+  /**
+   * Update a collection
+   * @param id - ID of the collection
+   * @param name - Name of the collection
+   * @param desscription - Description of the collection
+   * @returns
+   * Returns an updated Collection object
+   */
+  public async updateCollection(
+    id: string,
+    name: string,
+    description: string
+  ): Promise<Collection> {
+    const res = await this.vhttp.patch<CollectionResponse, object>(
+      [collection, id],
+      {
+        name,
+        description,
+      }
+    );
+    const convertedData = fromSnakeToCamel(res.data) as CollectionBase;
+    return new Collection(this.vhttp, convertedData);
   }
 
   /**